wip: started implementing typedskiplist trait
